# src/docs/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "CarKit API Documentation",
    "version": "0.0.1"
  },
  "host": [
    "carkit.lazyy.fr"
  ],
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/spending/{carId}": {
      "post": {
        "tags": [
          "Spending"
        ],
        "description": "Add a new spending entry for a car.",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "description": "Spending entry data",
            "schema": {
              "type": "object",
              "properties": {
                "cost": {
                  "example": 150.75
                },
                "date": {
                  "example": "2023-01-01"
                },
                "type": {
                  "example": "insurance"
                },
                "name": {
                  "example": "Annual insurance"
                }
              }
            }
          },
          {
            "in": "path",
            "name": "carId",
            "required": true,
            "description": "Path parameter carId",
            "schema": {
              "type": "string",
              "example": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Response",
            "examples": {
              "id": "uuid",
              "cost": 150.75,
              "date": "2023-01-01",
              "type": "insurance",
              "name": "Annual insurance"
            }
          },
          "401": {
            "description": "Response",
            "examples": {
              "status": 401,
              "msg": "Unauthorized."
            }
          },
          "404": {
            "description": "Response",
            "examples": {
              "status": 404,
              "msg": "Car not found."
            }
          }
        }
      },
      "get": {
        "tags": [
          "Spending"
        ],
        "description": "Get all spending entries for a car.",
        "parameters": [
          {
            "in": "path",
            "name": "carId",
            "required": true,
            "description": "Path parameter carId",
            "schema": {
              "type": "string",
              "example": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Response",
            "examples": [
              {
                "id": "uuid",
                "cost": 150.75,
                "date": "2023-01-01",
                "type": "insurance",
                "name": "Annual insurance"
              }
            ]
          },
          "401": {
            "description": "Response",
            "examples": {
              "status": 401,
              "msg": "Unauthorized."
            }
          },
          "404": {
            "description": "Response",
            "examples": {
              "status": 404,
              "msg": "Car not found."
            }
          }
        }
      }
    },
    "/spending/{id}": {
      "delete": {
        "tags": [
          "Spending"
        ],
        "description": "Delete a spending entry.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Path parameter id",
            "schema": {
              "type": "string",
              "example": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Response",
            "examples": {
              "status": 200,
              "msg": "Spending entry deleted successfully."
            }
          },
          "401": {
            "description": "Response",
            "examples": {
              "status": 401,
              "msg": "Unauthorized."
            }
          },
          "404": {
            "description": "Response",
            "examples": {
              "status": 404,
              "msg": "Spending entry not found."
            }
          }
        }
      }
    },
    "/service/{partId}": {
      "post": {
        "tags": [
          "Service"
        ],
        "description": "Create a new service entry for a part.",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "description": "Service creation data",
            "schema": {
              "type": "object",
              "properties": {
                "date": {
                  "example": "2023-01-01"
                },
                "mileage": {
                  "example": 15000
                }
              }
            }
          },
          {
            "in": "path",
            "name": "partId",
            "required": true,
            "description": "Path parameter partId",
            "schema": {
              "type": "string",
              "example": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Response",
            "examples": {
              "id": "uuid",
              "date": "2023-01-01",
              "mileage": 15000
            }
          },
          "401": {
            "description": "Response",
            "examples": {
              "status": 401,
              "msg": "Unauthorized."
            }
          },
          "404": {
            "description": "Response",
            "examples": {
              "status": 404,
              "msg": "Part not found."
            }
          }
        }
      },
      "get": {
        "tags": [
          "Service"
        ],
        "description": "Get all service entries for a part.",
        "parameters": [
          {
            "in": "path",
            "name": "partId",
            "required": true,
            "description": "Path parameter partId",
            "schema": {
              "type": "string",
              "example": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Response",
            "examples": [
              {
                "id": "uuid",
                "date": "2023-01-01",
                "mileage": 15000
              }
            ]
          },
          "401": {
            "description": "Response",
            "examples": {
              "status": 401,
              "msg": "Unauthorized."
            }
          },
          "404": {
            "description": "Response",
            "examples": {
              "status": 404,
              "msg": "Part not found."
            }
          }
        }
      }
    },
    "/service/{id}": {
      "put": {
        "tags": [
          "Service"
        ],
        "description": "Update a service entry.",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "description": "Service update data",
            "schema": {
              "type": "object",
              "properties": {
                "date": {
                  "example": "2023-02-01"
                },
                "mileage": {
                  "example": 16000
                }
              }
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Path parameter id",
            "schema": {
              "type": "string",
              "example": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Response",
            "examples": {
              "id": "uuid",
              "date": "2023-02-01",
              "mileage": 16000
            }
          },
          "401": {
            "description": "Response",
            "examples": {
              "status": 401,
              "msg": "Unauthorized."
            }
          },
          "404": {
            "description": "Response",
            "examples": {
              "status": 404,
              "msg": "Service not found."
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Service"
        ],
        "description": "Delete a service entry.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Path parameter id",
            "schema": {
              "type": "string",
              "example": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Response",
            "examples": {
              "status": 200,
              "msg": "Service deleted successfully."
            }
          },
          "401": {
            "description": "Response",
            "examples": {
              "status": 401,
              "msg": "Unauthorized."
            }
          },
          "404": {
            "description": "Response",
            "examples": {
              "status": 404,
              "msg": "Service not found."
            }
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter your Bearer token in the format **Bearer <token>**"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
...
}